Multisig escrow automation (#2)
* Added multisig and escrow tests

* Fix assert check

* Fixed multilevel multisig tests

* address comments

* address comments

* address comments

+++ integrationtests/src/test/java/io/nem/automation/example/ExampleSteps.java
@@ -145,7 +145,7 @@ public class ExampleSteps {
 		assertEquals(
 				testContext.getConfigFileReader().getCatCurrencyId().longValue(),
 				accountInfo.getMosaics().get(0).getId().getId().longValue());
-		assertEquals((long) transferAmount, accountInfo.getMosaics().get(0).getAmount().longValue());
+		assertEquals(transferAmount, accountInfo.getMosaics().get(0).getAmount().longValue());
 
 		// Verify the signer/sender account got update
 		AccountInfo signerAccountInfoBefore =
